Refactor electron main code
